# internal/server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Upload Assistant API",
    "version": "2.0.0-dev",
    "description": "Local-first Chinese PT retorrent automation with durable steps, rule gates, and auditable evidence."
  },
  "servers": [{"url": "http://127.0.0.1:8080"}],
  "security": [{"bearerAuth": []}],
  "paths": {
...
    "/api/v2/downloaders/{name}/probe": {
      "post": {"operationId": "probeDownloader", "parameters": [{"$ref": "#/components/parameters/ResourceName"}], "responses": {"200": {"description": "Authenticated downloader version probe", "content": {"application/json": {"schema": {"type": "object"}}}}, "502": {"$ref": "#/components/responses/Problem"}}}
    },
    "/api/v2/downloaders/{name}/snapshot": {
      "get": {
        "operationId": "getDownloaderSnapshot",
        "description": "Explicit read-only dashboard snapshot. Concurrent reads are coalesced and briefly cached. Torrent save paths, tracker URLs and credentials are excluded.",
        "parameters": [
          {"$ref": "#/components/parameters/ResourceName"},
          {"name": "filter", "in": "query", "schema": {"type": "string", "enum": ["all", "downloading", "seeding", "active", "paused", "checking", "error", "completed"], "default": "all"}},
          {"name": "query", "in": "query", "schema": {"type": "string", "maxLength": 200}},
          {"name": "offset", "in": "query", "schema": {"type": "integer", "minimum": 0, "default": 0}},
          {"name": "limit", "in": "query", "schema": {"type": "integer", "minimum": 1, "maximum": 200, "default": 100}}
        ],
        "responses": {"200": {"description": "Bounded live downloader snapshot", "content": {"application/json": {"schema": {"type": "object", "required": ["ok", "status", "snapshot", "blockers", "next_actions"], "properties": {"ok": {"type": "boolean", "const": true}, "status": {"type": "string", "const": "ready"}, "snapshot": {"$ref": "#/components/schemas/DownloaderDashboardSnapshot"}, "blockers": {"type": "array"}, "next_actions": {"type": "array"}}}}}}, "400": {"$ref": "#/components/responses/Problem"}, "502": {"$ref": "#/components/responses/Problem"}}
      }
    },
    "/api/v2/downloaders/{name}/torrents": {
      "post": {"operationId": "addDownloaderTorrent", "description": "Persists an add-intent audit before the remote write. Ambiguous or partial results return a blocked response with expected hash evidence and must be reconciled by read-only inspection before any further add.", "parameters": [{"$ref": "#/components/parameters/ResourceName"}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/AddTorrentRequest"}}}}, "responses": {"200": {"description": "Torrent added with hash evidence", "content": {"application/json": {"schema": {"type": "object"}}}}, "400": {"$ref": "#/components/responses/Problem"}, "409": {"description": "Downloader add requires reconciliation", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/DownloaderAddReconciliationResponse"}}}}, "502": {"$ref": "#/components/responses/Problem"}}}
    },
    "/api/v2/downloaders/{name}/torrents/{hash}": {
      "get": {"operationId": "inspectDownloaderTorrent", "parameters": [{"$ref": "#/components/parameters/ResourceName"}, {"$ref": "#/components/parameters/TorrentHash"}], "responses": {"200": {"description": "Torrent state and mapped path evidence", "content": {"application/json": {"schema": {"type": "object"}}}}, "404": {"$ref": "#/components/responses/Problem"}}}
    },
    "/api/v2/downloaders/{name}/torrents/{hash}/files": {
      "get": {
        "operationId": "listDownloaderTorrentFiles",
        "description": "Explicitly read one torrent's files. The response page is bounded even when the downloader returns a larger file set.",
        "parameters": [{"$ref": "#/components/parameters/ResourceName"}, {"$ref": "#/components/parameters/TorrentHash"}, {"name": "offset", "in": "query", "schema": {"type": "integer", "minimum": 0, "default": 0}}, {"name": "limit", "in": "query", "schema": {"type": "integer", "minimum": 1, "maximum": 500, "default": 100}}],
        "responses": {"200": {"description": "Bounded file evidence", "content": {"application/json": {"schema": {"type": "object"}}}}, "400": {"$ref": "#/components/responses/Problem"}, "404": {"$ref": "#/components/responses/Problem"}, "502": {"$ref": "#/components/responses/Problem"}}
      }
    },
    "/api/v2/downloaders/{name}/torrents/{hash}/limits": {
      "post": {"operationId": "setDownloaderTorrentLimits", "description": "Persists a limit-write intent before applying remote limits, then reads both limits back. Ambiguous writes, mismatched read-back, or result-audit failures return a blocked reconciliation response and must not be retried blindly.", "parameters": [{"$ref": "#/components/parameters/ResourceName"}, {"$ref": "#/components/parameters/TorrentHash"}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/SetTorrentLimitsRequest"}}}}, "responses": {"200": {"description": "Per-torrent limits applied and observed", "content": {"application/json": {"schema": {"type": "object"}}}}, "409": {"description": "Downloader limit write requires reconciliation", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/DownloaderLimitsReconciliationResponse"}}}}, "502": {"$ref": "#/components/responses/Problem"}}}
    },
...
    "/api/v2/image-hosts/{name}/probe": {
      "post": {
        "operationId": "probeImageHost",
        "description": "Explicit external-write probe that uploads one built-in 100x100 PNG compatible with normal thumbnail pipelines. The remote host may retain the image. The upload intent and bounded receipt evidence are audited and ambiguous outcomes are never retried automatically.",
        "parameters": [{"$ref": "#/components/parameters/ResourceName"}],
        "requestBody": {"required": true, "content": {"application/json": {"schema": {"type": "object", "additionalProperties": false, "required": ["confirm_upload"], "properties": {"confirm_upload": {"type": "boolean", "const": true}}}}}},
        "responses": {"200": {"description": "Test image uploaded and receipt validated"}, "400": {"$ref": "#/components/responses/Problem"}, "409": {"description": "Remote upload outcome is unknown"}, "502": {"$ref": "#/components/responses/Problem"}}
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {"type": "http", "scheme": "bearer", "bearerFormat": "ua_API_token"}
    },
    "parameters": {
...
      "ResourceName": {"name": "name", "in": "path", "required": true, "schema": {"type": "string", "pattern": "^[A-Za-z0-9][A-Za-z0-9._-]{0,63}$"}},
...
      "TorrentHash": {"name": "hash", "in": "path", "required": true, "schema": {"type": "string", "pattern": "^[a-fA-F0-9]{40}([a-fA-F0-9]{24})?$"}},
...
    },
    "responses": {
      "Problem": {"description": "Structured API problem", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Problem"}}}}
    },
    "schemas": {
...
      "Blocker": {
        "type": "object",
        "required": ["code", "message"],
        "properties": {"code": {"type": "string"}, "message": {"type": "string"}, "site_code": {"type": "string"}, "obligation_id": {"type": "string"}},
        "additionalProperties": true
      },
      "NextAction": {
        "type": "object",
        "required": ["action"],
        "properties": {"action": {"type": "string"}, "description": {"type": "string"}, "parameters": {"type": "object", "additionalProperties": true}},
        "additionalProperties": true
      },
...
	  "DownloaderAddReconciliationResponse": {
		"type": "object", "additionalProperties": false,
		"required": ["ok", "status", "error", "evidence", "blockers", "next_actions"],
		"properties": {
		  "ok": {"type": "boolean", "const": false},
		  "status": {"type": "string", "const": "blocked"},
		  "error": {"type": "object", "required": ["code", "detail"], "properties": {"code": {"type": "string", "enum": ["downloader_partial_add_requires_reconciliation", "downloader_add_outcome_unknown"]}, "detail": {"type": "string"}}},
		  "evidence": {"type": "object", "description": "Credential-free submitted torrent hashes, torrent SHA-256, adapter configuration SHA-256, and any trustworthy remote observation."},
		  "blockers": {"type": "array", "items": {"$ref": "#/components/schemas/Blocker"}},
		  "next_actions": {"type": "array", "items": {"$ref": "#/components/schemas/NextAction"}}
		}
	  },
	  "DownloaderLimitsReconciliationResponse": {
		"type": "object", "additionalProperties": false,
		"required": ["ok", "status", "error", "expected", "evidence", "blockers", "next_actions"],
		"properties": {
		  "ok": {"type": "boolean", "const": false},
		  "status": {"type": "string", "const": "blocked"},
		  "error": {"type": "object", "required": ["code", "detail"], "properties": {"code": {"type": "string", "const": "downloader_limits_outcome_unknown"}, "detail": {"type": "string"}}},
		  "expected": {"type": "object", "required": ["downloader_name", "torrent_hash", "download_limit", "upload_limit"], "properties": {"downloader_name": {"type": "string"}, "torrent_hash": {"type": "string"}, "download_limit": {"type": "integer", "minimum": 0}, "upload_limit": {"type": "integer", "minimum": 0}}},
		  "evidence": {"type": "object", "description": "Credential-free read-back evidence when available; it may be empty if the read-back itself failed."},
		  "blockers": {"type": "array", "items": {"$ref": "#/components/schemas/Blocker"}},
		  "next_actions": {"type": "array", "items": {"$ref": "#/components/schemas/NextAction"}}
		}
	  },
...
		"DownloaderDashboardSummary": {"type": "object", "additionalProperties": false, "required": ["total", "downloading", "seeding", "paused", "checking", "errors", "active", "download_speed", "upload_speed"], "properties": {"total": {"type": "integer", "minimum": 0}, "downloading": {"type": "integer", "minimum": 0}, "seeding": {"type": "integer", "minimum": 0}, "paused": {"type": "integer", "minimum": 0}, "checking": {"type": "integer", "minimum": 0}, "errors": {"type": "integer", "minimum": 0}, "active": {"type": "integer", "minimum": 0}, "download_speed": {"type": "integer", "minimum": 0}, "upload_speed": {"type": "integer", "minimum": 0}}},
		"DownloaderDashboardTorrent": {"type": "object", "additionalProperties": false, "required": ["hash", "name", "state", "state_group", "progress", "total_size", "amount_left", "downloaded", "uploaded", "download_speed", "upload_speed", "download_limit", "upload_limit", "limits_available", "ratio", "added_on", "completion_on", "time_active", "seeding_time"], "properties": {"hash": {"type": "string", "pattern": "^[a-fA-F0-9]{40}([a-fA-F0-9]{24})?$"}, "name": {"type": "string"}, "state": {"type": "string"}, "state_group": {"type": "string", "enum": ["downloading", "seeding", "paused", "checking", "error", "completed"]}, "progress": {"type": "number", "minimum": 0, "maximum": 1}, "total_size": {"type": "integer", "minimum": 0}, "amount_left": {"type": "integer", "minimum": 0}, "downloaded": {"type": "integer", "minimum": 0}, "uploaded": {"type": "integer", "minimum": 0}, "download_speed": {"type": "integer", "minimum": 0}, "upload_speed": {"type": "integer", "minimum": 0}, "download_limit": {"type": "integer", "minimum": 0}, "upload_limit": {"type": "integer", "minimum": 0}, "limits_available": {"type": "boolean", "description": "False when the list transport cannot safely read per-torrent limits without an unbounded number of remote calls."}, "ratio": {"type": "number", "minimum": 0}, "category": {"type": "string"}, "tags": {"type": "string"}, "added_on": {"type": "integer", "minimum": 0}, "completion_on": {"type": "integer", "minimum": 0}, "time_active": {"type": "integer", "minimum": 0}, "seeding_time": {"type": "integer", "minimum": 0}}},
		"DownloaderDashboardSnapshot": {"type": "object", "additionalProperties": false, "required": ["downloader_name", "adapter", "network_class", "fetched_at", "summary", "torrents", "filtered_total", "offset", "limit", "has_more"], "properties": {"downloader_name": {"type": "string"}, "adapter": {"type": "string", "enum": ["qbittorrent", "transmission", "deluge", "rtorrent"]}, "network_class": {"type": "string", "enum": ["unknown", "home", "seedbox"]}, "fetched_at": {"type": "string", "format": "date-time"}, "summary": {"$ref": "#/components/schemas/DownloaderDashboardSummary"}, "torrents": {"type": "array", "maxItems": 200, "items": {"$ref": "#/components/schemas/DownloaderDashboardTorrent"}}, "filtered_total": {"type": "integer", "minimum": 0}, "offset": {"type": "integer", "minimum": 0}, "limit": {"type": "integer", "minimum": 1, "maximum": 200}, "has_more": {"type": "boolean"}}},
...
      "AddTorrentRequest": {"type": "object", "additionalProperties": false, "required": ["torrent_base64"], "properties": {"torrent_base64": {"type": "string", "contentEncoding": "base64", "writeOnly": true}, "save_path": {"type": "string"}, "category": {"type": "string"}, "tags": {"type": "array", "items": {"type": "string"}}, "apply_labels": {"type": "boolean", "default": true, "description": "Must be explicitly false, with empty category/tags, when adapter capabilities do not support labels."}, "skip_checking": {"type": "boolean"}, "paused": {"type": "boolean"}, "upload_limit": {"type": "integer", "minimum": 0}, "download_limit": {"type": "integer", "minimum": 0}}},
      "SetTorrentLimitsRequest": {"type": "object", "additionalProperties": false, "required": ["download_limit", "upload_limit"], "properties": {"download_limit": {"type": "integer", "minimum": 0, "description": "Bytes per second; zero disables the limit."}, "upload_limit": {"type": "integer", "minimum": 0, "description": "Bytes per second; zero disables the limit."}}},
...
      "Problem": {"type": "object", "required": ["ok", "status", "error", "blockers", "next_actions"], "properties": {"ok": {"type": "boolean", "const": false}, "status": {"type": "string", "const": "failed"}, "error": {"type": "object", "required": ["code", "detail"], "properties": {"code": {"type": "string"}, "detail": {"type": "string"}}}, "blockers": {"type": "array", "items": {"$ref": "#/components/schemas/Blocker"}}, "next_actions": {"type": "array", "items": {"$ref": "#/components/schemas/NextAction"}}}}
    }
  }
}